requirements.txt
# Core dependencies
yfinance>=0.2.18
requests>=2.31.0
python-dotenv>=1.0.0

# Document Processing
pdfplumber>=0.9.0
python-docx>=0.8.11

# Text Processing
nltk>=3.8.1
regex>=2023.0.0

# Vector Embeddings
sentence-transformers>=2.2.0

# Vector Search
faiss-cpu>=1.7.4

# Technical Analysis - pandas-ta compatible versions
pandas-ta>=0.4.25b0
numpy==2.1.3
pandas==2.2.3
numba==0.61.0
llvmlite==0.44.0
python-dateutil==2.9.0.post0
pytz==2025.1
setuptools==75.8.0
six==1.17.0
tqdm==4.67.1
tzdata==2025.1

# Additional dependencies
scipy>=1.10.0

# Advanced Statistical Analysis
statsmodels>=0.14.0
scikit-learn>=1.3.0
arch>=6.2.0
